% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{cal_volcano_simple}
\alias{cal_volcano_simple}
\title{Compute linear model}
\usage{
cal_volcano_simple(data, colstart, colend, null_list, add_variable, like)
}
\arguments{
\item{data}{n_samples x n_features matrix}

\item{colstart}{feature start column number}

\item{colend}{feature end column number}

\item{null_list}{null model variables}

\item{add_variable}{additional variables}

\item{like}{likelihood ratio}
}
\value{
lm  linear model
}
\description{
Compute linear model
}
