% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_network.R
\name{correlation_network}
\alias{correlation_network}
\title{Making a correlation network}
\usage{
correlation_network(X, sel_features, options = list())
}
\arguments{
\item{X}{n_samples x n_features matrix}

\item{options}{options}

\item{y}{sel_features}
}
\value{
plot object
}
\description{
Making a correlation network
}
