% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_chord_diagram.R
\name{draw_cor_chord}
\alias{draw_cor_chord}
\title{Plot correlation as a chord diagram for given columns A and columns B}
\usage{
draw_cor_chord(df, feat.cols, title, pcut = 0.1, rcut = 0.5, colors = NULL)
}
\arguments{
\item{df:}{n_samples x n_features matrix}

\item{feat.cols:}{features group A, vector of strings}

\item{title:}{features group B, vector of strings}
}
\value{
none
}
\description{
Plot correlation as a chord diagram for given columns A and columns B
}
