% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{featureSelection}
\alias{featureSelection}
\title{export}
\usage{
featureSelection(
  X,
  y,
  method = "lasso",
  type = "classification",
  nFeatRep = 100,
  chooseS = "min",
  nLassoFolds = 5,
  thresh = 1,
  alpha = 1
)
}
\description{
export
}
