% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_chord_diagram.R
\name{gen_sec_color}
\alias{gen_sec_color}
\title{Generate palette for a chord diagram}
\usage{
gen_sec_color(col.feat, color_col = "assay", dlim = "_")
}
\arguments{
\item{col.feat:}{features with formatted names}
}
\value{
none
}
\description{
Generate palette for a chord diagram
}
