% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_correlation.R
\name{get_cor_col}
\alias{get_cor_col}
\title{Compute correlation for given columns A and columns B}
\usage{
get_cor_col(data, ab.labels, tcell.labels)
}
\arguments{
\item{data}{n_samples x n_features matrix}

\item{ab.labels}{features group A}

\item{tcell.labels}{features group B}
}
\value{
df data.frame with correlation and p value
}
\description{
Compute correlation for given columns A and columns B
}
