% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_correlation.R
\name{get_cor_tri}
\alias{get_cor_tri}
\title{Compute correlation among given columns A}
\usage{
get_cor_tri(data, from.labels, keep.tri = "U")
}
\arguments{
\item{data:}{n_samples x n_features matrix}

\item{from.labels:}{features group A, vector of strings}

\item{keep.tri:}{triangle region to keep}
}
\value{
df data.frame with correlation and p value
}
\description{
Compute correlation among given columns A
}
