% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelValidation.R
\name{modelValidation}
\alias{modelValidation}
\title{export}
\usage{
modelValidation(
  X,
  y,
  nFolds = 5,
  nReps = 10,
  nPerms = 100,
  nTop = 5,
  featureMethod = "lasso",
  method = "pls",
  type = "classification",
  nFeatRep = 100,
  nLassoFolds = 5,
  thresh = 1,
  alpha = 1,
  chooseS = "min",
  saveFlag = FALSE,
  fileStr = "accuraciesFullModel_withPerm"
)
}
\description{
export
}
