% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot linear model}
\usage{
plot_volcano(data, filename, title, foldername, pthreshold, col.arms)
}
\arguments{
\item{data}{n_samples x n_features matrix}

\item{filename}{str}

\item{title}{str}

\item{foldername}{str}

\item{pthreshold}{p value threshold}

\item{col.arms}{arms vector}
}
\value{
p  ploting object
}
\description{
Plot linear model
}
