% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_da.R
\name{pls_da}
\alias{pls_da}
\title{Partial Least Squares Discriminant Analysis (PLS-DA)
This function performs Partial Least Squares Discriminant Analysis on the given dataset.}
\usage{
pls_da(X, y, clr.grps, saved.dir, prefix, feature_selection_th = 0.8)
}
\arguments{
\item{X, }{matrix of predictive features, rows for samples and columns for measurements}

\item{y, }{vector of outcome variables}

\item{clr.grps, }{color palettes vector, with names corresponding to possible
outcomes}

\item{saved.dir, }{exporting directory/path}

\item{prefix, }{prefix to the file names}

\item{feature_selection_th, }{feature selection threshold (optional)}
}
\value{
None
}
\description{
Partial Least Squares Discriminant Analysis (PLS-DA)
This function performs Partial Least Squares Discriminant Analysis on the given dataset.
}
