% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_da.R
\name{pls_da_bar_stats}
\alias{pls_da_bar_stats}
\title{My customized plotting schemes for PLS-DA barplots
This function performs customized plotting for PLS-DA.}
\usage{
pls_da_bar_stats(model, options = list())
}
\arguments{
\item{model, }{trained ropls object}

\item{options, }{list of plotting options}
}
\value{
df_load, data.frame containing loading values for each latent dimension
}
\description{
My customized plotting schemes for PLS-DA barplots
This function performs customized plotting for PLS-DA.
}
