% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsBarplot.R
\name{plsBarplot}
\alias{plsBarplot}
\title{Visualization for PLS-R/PLS-DA results, namely, barplots for VIP scores and loadings of LV1}
\usage{
plsBarplot(
  oplsda,
  y,
  X,
  type = "classification",
  feature_annot = data.frame(),
  saveFlag = FALSE,
  fileStr = "",
  colors_bars = NA,
  width = 5,
  height = 4,
  vipFlag = FALSE,
  orth = TRUE,
  markEnrich = TRUE
)
}
\arguments{
\item{oplsda}{PLS object}

\item{y}{label/values}

\item{X}{data matrix (of selected features)}

\item{type}{"classification" or "regression"}

\item{feature_annot}{data frame with annotations for the features (colors,...)}

\item{saveFlag}{whether to save the figures}

\item{fileStr}{string where figures should be saved}

\item{colors_bars}{color for the classes}

\item{width}{of saved figure}

\item{height}{of saved figure}

\item{vipFlag}{whether only VIP scores > 1 should be plotted}

\item{orth}{flag whether orthogonalized version of PLS is used}

\item{markEnrich}{flag whether bars should be colored according to enrichment in class, only works for 2 classes}
}
\value{

}
\description{
Visualization for PLS-R/PLS-DA results, namely, barplots for VIP scores and loadings of LV1
}
