% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsBiplot.R
\name{plsBiplot}
\alias{plsBiplot}
\title{Visualization for PLS-R/PLS-DA results}
\usage{
plsBiplot(
  oplsda,
  y,
  type = "classification",
  feature_annot = data.frame(),
  saveFlag = FALSE,
  fileStr = "",
  alpha_loading = 1,
  color1 = "red",
  color2 = "blue",
  myColors_class = NA,
  shape1 = 21,
  shape2 = 21,
  fontsize = 4,
  orth = TRUE
)
}
\arguments{
\item{oplsda}{PLS object}

\item{y}{label/values}

\item{type}{"classification" or "regression"}

\item{feature_annot}{data frame with annotations for the features (colors,...)}

\item{saveFlag}{whether to save the figures}

\item{fileStr}{string where figures should be saved}

\item{alpha_loading}{value for plotting the loadings vectors}

\item{orth}{flag whether orthogonalized version of PLS is used}
}
\value{

}
\description{
Visualization for PLS-R/PLS-DA results
}
