% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ropls.R
\name{predict_ropls}
\alias{predict_ropls}
\title{Prediction using a ropls object}
\usage{
predict_ropls(model, X)
}
\arguments{
\item{model}{ropls object}

\item{X}{n_samples x n_features matrix}
}
\value{
vector of predicted labels
}
\description{
Prediction using a ropls object
}
