% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_to_asteriks.R
\name{pval_to_asteriks}
\alias{pval_to_asteriks}
\title{Transforms p-values into significance asteriks}
\usage{
pval_to_asteriks(pval)
}
\arguments{
\item{pval}{vector of p-values}
}
\value{
vector of strings with asteriks
}
\description{
Transforms p-values into significance asteriks
}
