% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roplsBiplot.R
\name{roplsBiplot}
\alias{roplsBiplot}
\title{Visualization for PLS-R/PLS-DA results, namely, score and loadings plot}
\usage{
roplsBiplot(
  ropls_obj,
  y,
  type = "classification",
  feature_annot = data.frame(),
  saveFlag = FALSE,
  fileStr = "",
  alpha_loading = 1,
  colors_y = NA,
  fontsize = 4,
  orth = TRUE
)
}
\arguments{
\item{ropls_obj}{ropls (O)PLS-DA object}

\item{y}{label/values}

\item{type}{"classification" or "regression"}

\item{feature_annot}{data frame with annotations for the features (colors,...)}

\item{saveFlag}{whether to save the figures}

\item{fileStr}{string where figures should be saved}

\item{alpha_loading}{value for plotting the loadings vectors}

\item{colors_y}{color for the classes}

\item{fontsize}{fontzise used for the plots}

\item{orth}{flag whether orthogonalized version of PLS is used}
}
\value{

}
\description{
Visualization for PLS-R/PLS-DA results, namely, score and loadings plot
}
