% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_scores.R
\name{score_accuracy}
\alias{score_accuracy}
\title{Accuracy}
\usage{
score_accuracy(y, y_pred)
}
\arguments{
\item{y}{true labels}

\item{y_pred}{predicted labels}
}
\value{
classification accuracy
}
\description{
Accuracy
}
