% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_scores.R
\name{score_mse}
\alias{score_mse}
\title{Mean squared error}
\usage{
score_mse(y, y_pred)
}
\arguments{
\item{y}{true labels}

\item{y_pred}{predicted labels}
}
\value{
mean squared error
}
\description{
Mean squared error
}
