% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_scores.R
\name{score_r2}
\alias{score_r2}
\title{R^2 (coefficient of determination)}
\usage{
score_r2(y, y_pred)
}
\arguments{
\item{y}{true labels}

\item{y_pred}{predicted labels}
}
\value{
R2 value
}
\description{
R^2 (coefficient of determination)
}
