% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ropls.R
\name{train_ropls}
\alias{train_ropls}
\title{Training the ropls model}
\usage{
train_ropls(X, y, options = list())
}
\arguments{
\item{X}{n_samples x n_features matrix}

\item{y}{vector of labels}
}
\value{
ropls object
}
\description{
Training the ropls model
}
