% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Model validation performing cross-validation and permutation tests.}
\usage{
validate(X, y, method, options)
}
\arguments{
\item{X}{n_samples x n_features matrix}

\item{y}{vector of labels}

\item{method}{method list with train, predict, score, and optionally select}

\item{options}{options list}
}
\value{
cv prediction, cv scores, rf scores, pt scores
}
\description{
Model validation performing cross-validation and permutation tests.
}
