% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_ropls.R
\name{visualize_ropls_loadings_bar}
\alias{visualize_ropls_loadings_bar}
\title{Title}
\usage{
visualize_ropls_loadings_bar(model, y, LV_ind = 1, options = list())
}
\arguments{
\item{ind_LV}{}
}
\value{

}
\description{
Title
}
