% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_ropls.R
\name{visualize_ropls_scores}
\alias{visualize_ropls_scores}
\title{Score plot}
\usage{
visualize_ropls_scores(model, y, options = list())
}
\arguments{
\item{model}{ropls object}

\item{y}{vector of labels}

\item{options}{}
}
\value{

}
\description{
Score plot
}
