% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeModelValidation.R
\name{visualizeModelValidation}
\alias{visualizeModelValidation}
\title{export}
\usage{
visualizeModelValidation(
  res,
  type,
  saveFlag = FALSE,
  strSave = "",
  modelColor = "gray"
)
}
\description{
export
}
