% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeRandomForest_RFE.R
\name{visualizeRandomForest_RFE}
\alias{visualizeRandomForest_RFE}
\title{Plots the out-of-bag error for recursive feature elimination (RFE) for random
forest}
\usage{
visualizeRandomForest_RFE(X, y)
}
\arguments{
\item{X}{data matrix}

\item{y}{label}
}
\description{
Plots the out-of-bag error for recursive feature elimination (RFE) for random
forest
}
