% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{featureSelection}
\alias{featureSelection}
\title{Performs feature selection using different methods.}
\usage{
featureSelection(
  X,
  y,
  method = "lasso",
  type = "classification",
  nFeatRep = 100,
  chooseS = "min",
  nLassoFolds = 5,
  thresh = 1,
  alpha = 1
)
}
\arguments{
\item{X}{data matrix, the column names are the featues}

\item{y}{label/values}

\item{method}{for feature selection:
\itemize{
 \item "none": all features are used
 \item "lasso": performs \code{nFeatRep} repetitions of penalized regression (linear for type = "regression", logistic for type = "classification")
 are performed and the features which are chosen at least \code{thresh} times are selected. Requires the package `glmnet`
 \item "lasso_min_mse": performs \code{nFeatRep} repetitions of penalized regression (linear for type = "regression", logistic for type = "classification")
 and choses the feature set with the overall lowest mean squared error (MSE).
 \item"randomForest_RFE": random forest recursive feature elimination (RFE). In each iteration the feature with lowest improtance is eliminated,
 finally the feature set is chosen which has the overall lowest out of bag error. Requires the package `randomForest`
 }}

\item{nFeatRep}{only used for method = "lasso" or "lasso_min_mse"}

\item{chooseS}{determines whether the regularization parameter is chosen for each repetition of LASSO which
provides the minimal MSE ("min"), or the most regularized model within one standard error of the minimal MSE is chosen.
Only used for method = "lasso" or "lasso_min_mse"}

\item{nLassoFolds}{Performs \code{nLassoFolds}-fold crossvalidation to determine penalization parameter. only used for method = "lasso"}

\item{thresh}{only used for method = "lasso"}

\item{alpha}{Elastic-net mixing parameter (see documentatino of glmnet). alpha = 1 corresponds to standard LASSO.
Only used for method = "lasso" or "lasso_min_mse"}
}
\value{
Names of the selected features
}
\description{
Performs feature selection using different methods.
}
