% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_correlation.R
\name{get_cor_col}
\alias{get_cor_col}
\title{Compute correlation for given columns A and columns B}
\usage{
get_cor_col(data, from.labels, to.labels)
}
\arguments{
\item{data:}{n_samples x n_features matrix}

\item{from.labels:}{features group A, vector of strings}

\item{to.labels:}{features group B, vector of strings}
}
\value{
df data.frame with correlation and p value
}
\description{
Compute correlation for given columns A and columns B
}
