% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelValidation.R
\name{modelValidation}
\alias{modelValidation}
\title{Model validation using permutation testing}
\usage{
modelValidation(
  X,
  y,
  nFolds = 5,
  nReps = 10,
  nPerms = 100,
  type = "classification",
  method = "pls",
  featureMethod = "lasso",
  nFeatRep = 100,
  nLassoFolds = 5,
  thresh = 1,
  alpha = 1,
  chooseS = "min",
  saveFlag = FALSE,
  fileStr = "accuraciesFullModel_withPerm",
  yPredOut = FALSE
)
}
\arguments{
\item{X}{data matrix}

\item{y}{label/values}

\item{nFolds}{for n-fold crossvalidatoin}

\item{nReps}{number of repetitions}

\item{nPerms}{number of permutations per repetition}

\item{type}{"classification" or "regression"}

\item{method}{for classification/regression, options are "pls", "randomForest", "logisticRegression",}

\item{featureMethod}{method for feature selection, @seealso \link{featureSelection.R}}

\item{nFeatRep}{@seealso \link{featureSelection.R}}

\item{nLassoFolds}{@seealso \link{featureSelection.R}}

\item{thresh}{@seealso \link{featureSelection.R}}

\item{alpha}{@seealsoe \link{featureSelection.R}}

\item{chooseS}{@seealso \link{featureSelection.R}}

\item{saveFlag}{whether to save the output}

\item{fileStr}{string where RData should be saved}

\item{yPredOut}{flag whether predicted output should be returned}
}
\value{
Metrics of performance (accuracy or correlation/rmses) for models based on
real data or random features/ ermuted labels
}
\description{
Model validation using permutation testing
}
