% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_denoising.R
\name{multilevel_denoising}
\alias{multilevel_denoising}
\title{Multilevel Denoising}
\usage{
multilevel_denoising(X, X_labels)
}
\arguments{
\item{X}{raw data matrix}

\item{X_labels}{vector/factor indicating subject id}
}
\value{
denoised and z-scored data matrix
}
\description{
Multilevel Denoising
}
