% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_da.R
\name{my_visualize_ropls_loadings_bar}
\alias{my_visualize_ropls_loadings_bar}
\title{My customized plotting schemes for PLS-DA barplots
This function performs customized plotting for PLS-DA.}
\usage{
my_visualize_ropls_loadings_bar(model, options = list())
}
\arguments{
\item{model, }{trained ropls object}

\item{options, }{list of plotting options}
}
\value{
a list [bar, df_load], bar is ggplot handle and data.frame containing loading values for each latent
dimension
}
\description{
My customized plotting schemes for PLS-DA barplots
This function performs customized plotting for PLS-DA.
}
