% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsBarplot.R
\name{plsBarplot}
\alias{plsBarplot}
\title{export}
\usage{
plsBarplot(
  oplsda,
  y,
  X_sel,
  saveFlag = FALSE,
  fileStr = "",
  color1 = "red",
  color2 = "blue",
  feature_annot = data.frame()
)
}
\description{
export
}
