% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsBiplot.R
\name{plsBiplot}
\alias{plsBiplot}
\title{export}
\usage{
plsBiplot(
  oplsda,
  y,
  saveFlag = FALSE,
  fileStr = "",
  color1 = "red",
  color2 = "blue",
  shape1 = 21,
  shape2 = 21,
  fontsize = 4,
  alpha_loading = 1,
  feature_annot = data.frame()
)
}
\description{
export
}
