% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_repeat.R
\name{select_repeat}
\alias{select_repeat}
\title{Repeat the feature selection and return features that are selected more
often than a provided threshold}
\usage{
select_repeat(X, y, selector, options = list())
}
\arguments{
\item{X}{n_samples x n_features matrix}

\item{y}{vector of labels}

\item{selector}{feature selection method}

\item{options}{options list}
}
\value{

}
\description{
Repeat the feature selection and return features that are selected more
often than a provided threshold
}
