% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_ropls.R
\name{visualize_ropls_loadings_bar}
\alias{visualize_ropls_loadings_bar}
\title{Barplot for loadings}
\usage{
visualize_ropls_loadings_bar(model, options = list())
}
\arguments{
\item{model}{ropls object}

\item{ind_LV}{latent variable index}
}
\value{

}
\description{
Barplot for loadings
}
