% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_validate.R
\name{visualize_validate}
\alias{visualize_validate}
\title{Visualization of validation results}
\usage{
visualize_validate(vals, options = list())
}
\arguments{
\item{options}{}
}
\value{
plot handle
}
\description{
Visualization of validation results
}
