% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeModelValidation.R
\name{visualizeModelValidation}
\alias{visualizeModelValidation}
\title{Generates figures for the model validation using permutation tests and returns the p-values}
\usage{
visualizeModelValidation(
  res,
  type,
  saveFlag = FALSE,
  fileStrPart = "modelValidation.pdf",
  modelColor = "gray"
)
}
\arguments{
\item{res}{result file obtained from modelValidation}

\item{type}{"classification" or "regression"}

\item{saveFlag}{TRUE or FALSE}

\item{fileStrPart}{string where figures should be saved,
will be extended with "_acc" for classification or "_corr" and "_rmses" for regression problems}

\item{modelColor}{color for model violin plot}
}
\value{
p-values for model
}
\description{
Generates figures for the model validation using permutation tests and returns the p-values
}
